Re-organises src again.
