+++ .rubocop.yml
@@ -37,8 +37,8 @@ Metrics/MethodLength:
 
 Naming/AccessorMethodName:
   Exclude:
-    - 'lib/starter/templates/files.rb'
-    - 'lib/starter/templates/endpoints.rb'
+    - 'lib/starter/builder/files.rb'
+    - 'lib/starter/builder/endpoints.rb'
 
 Style/AsciiComments:
   Enabled: false
